Define Chip8Sound interface
So it's mockable and swappable.
